backend/requirements.txt
bottle
mysql-connector-python
requests